+++ src/Exception/Fixer/EncodingException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EasyAudit\Exception\Fixer;
+
+class EncodingException extends \RuntimeException
+{
+
+}