+++ src/Exception/Scanner/FunctionNotFoundException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EasyAudit\Exception\Scanner;
+
+class FunctionNotFoundException extends \RuntimeException
+{
+
+}